# x-pack/filebeat/input/cel: use structpb.Struct as intermediate type